@@ -49,8 +49,8 @@ return [
             'host' => env('DB_HOST', '127.0.0.1'),
             'port' => env('DB_PORT', '3306'),
             'database' => env('DB_DATABASE', 'u383320855_lGseh'),
-            'username' => env('DB_USERNAME', 'u383320855_GRm2Z'),
-            'password' => env('DB_PASSWORD', '0123456789+-*Ou'),
+            'username' => env('DB_USERNAME', 'root'),
+            'password' => env('DB_PASSWORD', ''),
             'unix_socket' => env('DB_SOCKET', ''),
             'charset' => 'utf8mb4',
             'collation' => 'utf8mb4_unicode_ci',
